Fix type annotations to run with Python 3.9
According to the `tool.poetry.dependencies.python` key of
`pyproject.toml`, the package supports `">=3.9,<3.13"`. However, when
running `r2r --help` with Python 3.9, I received the following error:

> `TypeError: unsupported operand type(s) for |: 'type' and 'NoneType'`

The problem was that one set of type annotations uses the `|` operator
instead of `typing.Union`, which was not supported until Python 3.10. I
switched to using the `typing` module to restore support for Python 3.9.