Feature/merge dev and main (#967)
* Fix argument order on ingest_files (#883)

* change get basename to get abspath of config (#893)

* Fix typos (#919)

* Fix typo in README (`Pyhon` -> `Python`)

* Fix incorrect setting name in error message

`collection` was used, but it should be `vecs_collection`.

* Fix type annotations to run with Python 3.9 (#943)

According to the `tool.poetry.dependencies.python` key of
`pyproject.toml`, the package supports `">=3.9,<3.13"`. However, when
running `r2r --help` with Python 3.9, I received the following error:

> `TypeError: unsupported operand type(s) for |: 'type' and 'NoneType'`

The problem was that one set of type annotations uses the `|` operator
instead of `typing.Union`, which was not supported until Python 3.10. I
switched to using the `typing` module to restore support for Python 3.9.

* Fix typo in variable name (#961)

* Add support for relative paths for the prompt directory in config (#958)

---------

Co-authored-by: Tim Kellogg <timothy.kellogg@gmail.com>
Co-authored-by: Maxine Lai <76145646+mmtmr@users.noreply.github.com>
Co-authored-by: Alex Hedges <aphedges@users.noreply.github.com>
Co-authored-by: Manuel R. Ciosici <manuelrciosici@gmail.com>