Allow api_base/api_key in embedding config (#2269)
Currently, to have embedding and completion hit two distinct
API base URLs, you have to define the embedding base in the
environment variable and set the completion base in the toml
config. This patch obviates the need to use the environment
at all and allows setting api_base and api_key in the config
for embedding, as well.

Co-authored-by: Clay Fouts <cfouts@ptfs.com>